Changed the stack interface to not use unsafe, changed the ExternFunction type
Replaced the UnsafeCell with a RefCell to make the stack safe to use. The RefCell is borrowed during vm execution and is then dropped during external function execution. Changed the type of ExternFunction to only take the VM, stack operations is done directly on the vm to make the interfacing easier (especially for C interop since only Rust understands dropping the StackFrame after the call)