fix(check::typecheck): Allow unification on partially applied aliases
This fix allows the following type alias to unify even when it is only partially applied.
```
type Test a b c = Type (a Int) b c

Test a b == Type (a Int) b
Test a == Type (a Int)
Test == ??? (Impossible to do a sane substitution)
```

The downside of this fix is that anytime two aliases are unified we do a full search through the aliases so when unifying two types which are deeply aliased unification may take a long time. A more efficient approach may use a disjoint-set to deduce wheter it is possible for two aliases to be equal.

type Test1 = {}
and Test2 = Test1
and Test3 = Test2

Unify Test2 <=> Test3
This causes Test2 => Test1 => {} on the lhs followed by Test3 => Test2 => Test1 => {} (a search through 5 aliases which is unnecessarily expensive).