feat(vm): Allow values on the stack to be rooted
Introduced the ability to lock the stack as it is at a certain point allowing every pointer below that point to be assumed as rooted. This in turns mean that it should be safe to take references to the innards of those pointers which used in the previous commit. This could probably be done in a safer way but this is still an improvement to the current state of things.