refactor: Modify the Type enum to remove the TypeConstructor type
The TypeConstructor enum was always in an awkward position as it limited what could be put in the Data variant. This change allows any type to be put in that slot which makes the typechecking code slightly more awkward but every other part of the code is cleaner instead. The typechecking code should likely be possible to improved as well.