refactor: Change functions operating on VM to just operator on Thread
This change necessitates that `Thread` is always allocated on the garbage collacted heap as the current thread will need to be mark at the start of a collection. As a consequence of this change VM is now just a wrapper around Thread which only makes sure that a thread is traversed while the VM wrapper is alive.