feat(check): Store alias data directly in the type when possible
This uses the Type::Alias variant to store information about the number of arguments and the type which is aliased directly in the `Type` enum. This should eventually make it possible to always retrive the aliased type directly without needing to bring it into scope which will for instance allow `vm.get_global("prelude.eq_Int.(+)` to succeed.