fix: Use the alias information stored inside an alias type to resolve type aliases
This utilizes the `Type::Alias` variant whenever it is possible to directly resolve the alias instead of looking it up through a `TypeEnv`. In addition to being faster this also allows instantiation of alias types even when the type is not in scope which makes certain operation much easier to implement.