feat(vm): Reduce the memory needed for arrays in the vm
Adds the ValueArray type which keeps track of the type of values it contains in a header instead of storing the information next to each value. Not only does this reduce the memory needed for the current arrays by half but if smaller primitive types are added like a byte type the reduction will be by a factor of 16.