fix: Do not traverse the replacement type in walk_move_type
When using walk_move_type to instantiate a type such as `State s (a -> b)` where `type State s a = s -> { state: s, value: a }` the default behaviour of walking the replacement type would cause a stack overflow due to:

Instantiate: a ==> (a -> b)
Walk: (a -> b)
Instantiate: a ==> (a -> b)
Walk: (a -> b)
etc.

There are one case when walking the replacement is desired but that is easily fixed by recursively walking the replacement before returning it (which should be more efficient as well).