fix(parser): Update combine-language to v2.0.0-beta2
This gives access to parsers which do not skip whitespace after parsing the actual token which fixes the lexer not returning the correct spans. There are still a few expressions which do not have entirely correct spans but since the parser is getting rewritten with LALRPOP this works well enough for now.