Delay string interning until parse time
The strings are already allocated by the parser, so it is easier just to delay this until we hit the parser. This greatle simplifies the type signatures in the lexer, and will make it much easier to transition to LALRPOP. For efficiency's sake, we should eventually transition to using slices in the lexer though.