feat(vm): Implement field access of polymorphic records
This adds an extra field to the `Context` type which contains mappings for each record type allocate in the current thread. These mappings are used to calculate the offset for polymorphic field accesses. The current implementation uses a bit to much indirection for my liking as well as being rather expensive when it comes to allocating new records as it needs to check if a `Vec<InternedStr>` already exists in the field map. This can be changed later however if it is deemed to slow