fix: Replace the Generic<A> hack with type holes
`Generic<A>` was used as a way of indicating 'any value' would be acceptable. It was however a big hack however as it relied on the generic variable `a` being instantiated before unifying the type. This PR fixes that behaviour making `Generic<A>` instead act as a type signature. This breaks a lot of code which relied on this behaviour though so the `Hole` marker type was added so that specifying that the type is unknown/does not matter is still possible.