feat: Only store a function pointer for extern functions
This makes it possible to deep clone objects containing extern functions making it possible to send them up the thread tree. It also makes the C ffi more efficient as C functions can be used directly instead of needing to be wrapped in a Box<Fn(..)>.

The only downside is that closures can no longer be passed directly but this not used at the moment and could be re-added by adding an optional data pointer to `ExternFunction`.

The compile-fail tests unfortunately ended up less readable due to and issue with macros which causes the compiletest crate to use the wrong spans.