fix: Specify the libgluon dependency explicitly in compiletest.rs
Apparently cargo/rustc no longer removes .rlibs compiled with a different set of feature flags (or rustc are no longer capable of disambiguate between them). Since gluon are compiled with a combination of `test`, `nightly` and `skeptic` we end up with two libgluon-xxxxxxx.rlib in deps and I select just one of them in this PR which should hopefully be the correct one.