fix(compiler): Pop the correct amount of variables in the compiler
See compiler.rs:919-929. By using `Split` all fields in the value are pushed but when `Compiler::pop_pattern` is called to remove these variables it only removes the number of fields in the pattern (which may be fewer). To fix this I refactored the code a bit to use `ScopedMap` which is already used to do this kind of cleanup.