feat(parser): Return valid a valid AST after parsing an invalid expression
This adds a blanket error recovery in the grammar in the expression production which allows the parser to recover in a wide range of scenarios with reasonable accuracy. However it might be a bit to much of a catch-all as it does mean worse error messages in some cases such as the test `wrong_indent_expression` which now produces three errors instead of just one which it did before.