feat(vm): Allow extern functions to return futures
This integrates the futures crate which provides the functinality for
running asynchronous computations which is something that was almost
possible already thanks to the vm's ability to yield.

To make it explicit which functions are (potentially) asynchronous the
`Error::Yield` variant was replaced with `futures::Async`.
`AsyncPushable` were added to represent a subset of pushable values
which can be asynchronous. This is then used to only allow the direct
return value of a function to be asynchronous (`() -> Future<A>` works
but `() -> { x: Future<A> }` will not compile).