fix(repl): Don't panic when attempting to quit the repl
A bug in `catch_io` caused the call to the first io action to keep executing the code above the `catch_io` code instead of returning control to the caller of `catch_io`. By adding a lock, the call to the io action will cause `Thread::execute` to return control back when it encounters it so that the caller can continue executing the gluon code.