fix(check): Require that an record alias matches exactly on record construction
When creating a record the compiler attempts to look up a matching alias to simplify typechecking later on (since an alias is a smaller type than a record, most of the time). For the alias to unify correctly though its fields need to match exactly however whereas it used to only require a subset of them to match which is the case for looking up an alias from a field access or pattern match.