feat(vm): Let primitive modules be loaded via import!
This rewrites how primitive modules are loaded. Instead of immediately storing these modules in globals when `load` is called, the load function are now stored in the `import!` macro and called only when `import! <primitive module name>` is called in a gluon script.

This abstracts away `define_global` as well as global variables from users of gluon. Making it more obvious what bindings are in scope and makes primitive modules as first class as normal modules are in gluon.

The actual removal of global variables from the user perspective comes in a later commit.