fix(vm): Handle polymorphic records created in parent threads
Since the field_map (that is now removed) were only filled in when
records were actually constructed and they were associated to each
specific thread there would be crashes when records were used in
polymorphic functions.

This associated the field map to the already existing `TypeInfo` that is
stored in all gluon data so it can be used from any thread.

There is still some optimization that could be done to reduce the size
of the type info as the field map is unused in non-records and userdata.
It is not crucual though so can be fixed at any time later.