fix(vm): Don't let rustc think the Array type can't exist
THe use of `Void` in the `Array` type was meant to absolutely prevent
the `Array` type from existing except behind a pointer which is invalid
(except the empty array I guess) as the elements are stored inline
behind the length field.

Unfortunately https://github.com/rust-lang/rust/pull/45225 made rustc
think that fields of the `Array` type could not exist at all and placed
fields of `Array` at the start of any struct containing them, causing
any write to the `Array` to overwrite other fields.