chore: Run the gluon script tests without a test harness
This makes sure that the success/failure of each test are printed as they happen but forces us to declare all tests in the Cargo.toml instead. To make sure that we do not forget to do this I added a step to build.rs which does a sanity check against this to make sure this is at least caught in travis.