fix(rename): Rename implicit imports
This caused panics on linux but not on windows since it happened that
two implicit imports, one from the implicit prelude and one from the
normal code, laid on the exact same location and got the same name. It
didn't happen on windows since I have CRLF linendings there (but it
could just as well have happened on windows and not on linux of course).
Since these bindings were never renamed they looked the same to the
compiler which then loaded the wrong variable (since the implicit
prelude essentially were shadowed).