feat: Obey the new @infix attribute for operator fixity
Rather than the hardcoded table of associativity and precedence we now
look at the `@infix <FIXITY> <PRECEDENCE>` attribute of any defined
operator. This makes it possible to supply this information for
arbitrary user defined operators but it has the slight drawback in that
there are cases where this information does not exist.

```f#
let string = import! std.string
// `append` does not have the `@infix` attribute so it does not exist
for
// `++` either
let (++) = string.semigroup.append
```