feat: Return an error when infix operators are used without fixity
Due to the way metadata propagation works it is quite easy to use an
infix operator which has lost or otherwise do not have fixity specified.
Since this can give some rather confusing errors it is better to just
reject the code and force the fixity to be explicitly specified instead
of falling back to a default.