feat: Use #[IDENTIFIER(..)] as the attribute syntax
Changes the syntax of attributes by breaking them out of comments and
into their own syntactical element.

With the attribute in comment syntax it wasn't clear that attributes had
an effect on the code. The same syntax as Rust was chosen mostly out of
familiarity, if a compelling argument for another syntax is made it
could be changed (and the formatter could perhaps help in migration as
that should be simple).

Closes #515