Add missing newline after usage when no arguments given
When no arguments are given to `gluon_repl`, we print usage for the
command, e.g.:

    $ gluon
    USAGE:
        gluon [FLAGS] [OPTIONS] [FILE]... [SUBCOMMAND]

However, we were not printing a newline after the final character of
this usage string.  This is unsightly as it results in the shell
printing its `$PS1` immediately after the usage string without a
preceding line break.

With this commit, we now print a trailing newline after the usage
string.