Add missing newline when input file not found
When `gluon_repl` is given as an argument an input file that cannot be
found, or some other error occurs other than a VM error with a
message, we print the error using `gluon::Error::emit`.  E.g.:

    $ gluon file-not-there
    Could not find module 'file-not-there'. Searched `.`.

The `gluon::Error::emit` function does not print a trailing newline.
As `gluon_repl` was not adding this newline, we were not printing a
trailing newline after the error.  This is unsightly as it results in
the shell printing its `$PS1` immediately after the error without a
preceding line break.

With this commit, we now print a trailing newline after the error.