feat(check): Remove the alias dependence for variants
Before a variant could only be defined at the top level of a type definition since they needed to construct an instance of that type. This lifts that restriction and variant constructors now returns a row type of `forall a . | VARIANT_NAME ARGS*  | a`. Which ends up meaning that they construct an open variant row which contains at least the constructor in question, but could also contain more constructors.