feat(vm): Allow references to be passed through
Adds the `Ref` type which ensures that the borrowed reference is alive
as long as any outstanding references to it exist. Care must be taken
that any references to the object in gluon does not outlive the scope in
which `Ref` exists for or a gluon panic will occur if it is accessed
after `Ref` has gone out of scope.