feat: Automatically unroot values stored in `Gc` allocated values
`RootedThread` and `RootedValue` are reference counted pointers so
storing them in a value allocated on the GC heap will create a cycle
which will never be dropped.

To avoid this we detect when such a value is allocated on the GC heap
and remove the value from the root list.

Fixes #746