refactor: Make `Trace` an unsafe trait
Since `Trace` now manages wheter a value is rooted or not it is possible
to unroot a value, causing it to be collected while there are still
references to it.

Normally `Trace` will be derived in which case the implementation is
automatically correct since it will only call the correct method in each
of the methods.

BREAKING CHANGE