perf(check): Only add implicit fields if the binding is implicit
When encountering a binding we no longer recurse through the entire
record but instead only do it as long as we are in a implicit binding
already. Thus adding `Applicative` also adds its `Functor` field since
`Applicative` is implicit. However, for a normal record such as from
`let int = import! std.int` we no longer recurse and add all the
implicit instances in `std.int`, instead a `let int @ { ? }` is
required.

BREAKING CHANGE